Extract IJ specific options to own file
